Add support for configuring operation parameters without instantiation
This allows testing init parameters for feasibility by only running the init callback function and returning its success without creating a new operation.  Future operation instantiations can use additional configuration values stored in the OP_INIT_CONFIG subtree written during this init function call.  This is useful if the setup is computationally expensive and should be done elsewhere in advance.

Special init parameter key names are now constants.

Parameter KeyValues no longer rewind at the start of calls due to the possibility that OP_INIT_PARAM may exist as a subtree within the parameter KeyValues with the position set up in advance for the operation init function to read.  Consequently, parameter KeyValues used for operation instantiations must now be set up at the correct tree depth in advance, and init functions must restore the KeyValues to the original depth after the call in order for it to be reusable.