Add an optional cleanup function for LocalDataPack
This allows storing dynamically created objects within the edge data LocalDataPack provided in the cost function callback.  The implementation of the cleanup function should ensure all dynamically created data are properly freed to prevent memory leaks.

Edge data not returned as part of the path result of the search will call this cleanup function prior to being automatically destroyed.  Calling LocalDataPack.Destroy() on the edge data LocalDataPack returned in the result path will also automatically call this cleanup function.  The cleanup function is set for the LocalDataPack upon instantiation and cannot be changed.